requirements.txt
boto3==1.18.*
psycopg2-binary
cryptography